requirements.txt
streamlit
pandas
numpy
altair
pyserial